+++ gg-pingpong-api/src/main/java/gg/pingpong/api/admin/match/dto/MatchUser.java
@@ -1,0 +1,20 @@
+package gg.pingpong.api.admin.match.dto;
+
+import gg.pingpong.data.match.type.Option;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MatchUser {
+	private Long userId;
+	private String intraId;
+	private Option mode;
+
+	public MatchUser(Long userId, String intraId, Option mode) {
+		this.userId = userId;
+		this.intraId = intraId;
+		this.mode = mode;
+	}
+}
